create ClassRooms CRUD with Filter(checkBox)

+++ app/Http/Controllers/Grades/GradeController.php
@@ -79,22 +79,24 @@ class GradeController extends Controller
 
 
     public function destroy(Request $request)
+        # pulck --> make array for sort many ids
     {
-//        $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
+        $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
 
-//        if($MyClass_id->count() == 0){
+        if($MyClass_id->count() == 0){
 
-        $Grades = Grade::findOrFail($request->id)->delete();
-        toastr()->error(trans('messages.Delete'));
-        return redirect()->route('Grades.index');
-//        }
+            $Grades = Grade::findOrFail($request->id)->delete();
+            toastr()->error(trans('messages.Delete'));
+            return redirect()->route('Grades.index');
+        }
 
-//        else{
-//
-//            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
-//            return redirect()->route('Grades.index');
-//
-//        }
+        else{
+
+            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+            return redirect()->route('Grades.index');
+
+        }
+
 
     }
 }
